fix: statistic game1 in Statistics.ts
